Move code to structure::validator
