+++ components/support/nimbus-fml/src/intermediate_representation.rs
@@ -5,6 +5,7 @@ use crate::defaults::{DefaultsMerger, DefaultsValidator};
 use crate::error::FMLError::InvalidFeatureError;
 use crate::error::{FMLError, Result};
 use crate::frontend::{AboutBlock, FeatureMetadata};
+use crate::structure::StructureValidator;
 use crate::util::loaders::FilePath;
 use anyhow::{bail, Error, Result as AnyhowResult};
 use serde::{Deserialize, Serialize};
@@ -300,26 +301,8 @@ impl FeatureManifest {
     }
 
     pub fn validate_manifest(&self) -> Result<()> {
-        // We first validate that each enum_def has a unique name.
-        // TODO: We repeat this check three times, it should be its
-        // own generic helper
-        let mut enum_names = HashSet::new();
-        self.validate_enum_defs(&mut enum_names)?;
-        // We then validate that each obj_defs also has a unique name.
-        let mut obj_names = HashSet::new();
-        self.validate_obj_defs(&mut obj_names)?;
-
-        // We then validate that each feature_def has a unique name.
-        let mut feature_names = HashSet::new();
-        self.validate_feature_defs(&mut feature_names)?;
-
         // We then validate that each type_ref is valid
-        for feature_def in self.iter_feature_defs() {
-            for prop in &feature_def.props {
-                let path = format!("features/{}.{}", &feature_def.name, &prop.name);
-                Self::validate_type_ref(&path, &prop.typ, &enum_names, &obj_names)?;
-            }
-        }
+        self.validate_structure()?;
         self.validate_defaults()?;
 
         // Validating the imported manifests.
@@ -331,133 +314,13 @@ impl FeatureManifest {
         Ok(())
     }
 
-    fn validate_type_ref(
-        path: &str,
-        type_ref: &TypeRef,
-        enum_names: &HashSet<String>,
-        obj_names: &HashSet<String>,
-    ) -> Result<()> {
-        match type_ref {
-            TypeRef::Enum(name) => {
-                if !enum_names.contains(name) {
-                    return Err(FMLError::ValidationError(
-                        path.to_string(),
-                        format!(
-                            "Found enum reference with name: {}, but no definition",
-                            name
-                        ),
-                    ));
-                }
-                Ok(())
-            }
-            TypeRef::Object(name) => {
-                if !obj_names.contains(name) {
-                    return Err(FMLError::ValidationError(
-                        path.to_string(),
-                        format!(
-                            "Found object reference with name: {}, but no definition",
-                            name
-                        ),
-                    ));
-                }
-                Ok(())
-            }
-            TypeRef::EnumMap(key_type, value_type) => match key_type.as_ref() {
-                TypeRef::StringAlias(_) | TypeRef::Enum(_) => {
-                    Self::validate_type_ref(path, key_type, enum_names, obj_names)?;
-                    Self::validate_type_ref(path, value_type, enum_names, obj_names)
-                }
-                _ => Err(FMLError::ValidationError(
-                    path.to_string(),
-                    format!(
-                        "Map key must be a String, string-alias or enum, found: {:?}",
-                        key_type
-                    ),
-                )),
-            },
-            TypeRef::List(list_type) => {
-                Self::validate_type_ref(path, list_type, enum_names, obj_names)
-            }
-            TypeRef::StringMap(value_type) => {
-                Self::validate_type_ref(path, value_type, enum_names, obj_names)
-            }
-            TypeRef::Option(option_type) => {
-                if let TypeRef::Option(_) = option_type.as_ref() {
-                    Err(FMLError::ValidationError(
-                        path.to_string(),
-                        "Found nested optional types".into(),
-                    ))
-                } else {
-                    Self::validate_type_ref(path, option_type, enum_names, obj_names)
-                }
-            }
-            _ => Ok(()),
+    fn validate_structure(&self) -> Result<(), FMLError> {
+        let validator = StructureValidator::new(&self.enum_defs, &self.obj_defs);
+        for object in self.iter_object_defs() {
+            validator.validate_object_def(object)?;
         }
-    }
-
-    fn validate_enum_defs(&self, enum_names: &mut HashSet<String>) -> Result<()> {
-        for nm in self.enum_defs.keys() {
-            if !enum_names.insert(nm.clone()) {
-                return Err(FMLError::ValidationError(
-                    format!("enums/{}", nm),
-                    format!(
-                        "EnumDef names must be unique. Found two EnumDefs with the same name: {}",
-                        nm
-                    ),
-                ));
-            }
-        }
-        Ok(())
-    }
-
-    fn validate_obj_defs(&self, obj_names: &mut HashSet<String>) -> Result<()> {
-        for nm in self.obj_defs.keys() {
-            if !obj_names.insert(nm.clone()) {
-                return Err(FMLError::ValidationError(
-                    format!("objects/{}", nm),
-                    format!(
-                    "ObjectDef names must be unique. Found two ObjectDefs with the same name: {}",
-                    nm
-                ),
-                ));
-            }
-        }
-        Ok(())
-    }
-
-    fn validate_feature_defs(&self, feature_names: &mut HashSet<String>) -> Result<()> {
         for feature_def in self.iter_feature_defs() {
-            if !feature_names.insert(feature_def.name.clone()) {
-                return Err(FMLError::ValidationError(
-                    feature_def.name(),
-                    format!(
-                    "FeatureDef names must be unique. Found two FeatureDefs with the same name: {}",
-                    feature_def.name
-                ),
-                ));
-            }
-            // while checking the feature, we also check that each prop is unique within a feature
-            let mut prop_names = HashSet::new();
-            self.validate_props(feature_def, &mut prop_names)?;
-        }
-        Ok(())
-    }
-
-    fn validate_props(
-        &self,
-        feature_def: &FeatureDef,
-        prop_names: &mut HashSet<String>,
-    ) -> Result<()> {
-        let path = format!("features/{}", &feature_def.name);
-        for prop in &feature_def.props {
-            if !prop_names.insert(prop.name.clone()) {
-                return Err(FMLError::ValidationError(
-                    format!("{}.{}", path, prop.name),
-                    format!(
-                    "PropDef names must be unique. Found two PropDefs with the same name: {} in the same feature_def: {}",
-                    prop.name, feature_def.name
-                )));
-            }
+            validator.validate_feature_def(feature_def)?;
         }
         Ok(())
     }
@@ -468,90 +331,7 @@ impl FeatureManifest {
             validator.validate_object_def(object)?;
         }
         for feature in self.iter_feature_defs() {
-            self.validate_feature_structure(feature)?;
             validator.validate_feature_def(feature)?;
-        }
-        Ok(())
-    }
-
-    fn validate_feature_structure(&self, feature_def: &FeatureDef) -> Result<()> {
-        let mut string_aliases: HashSet<_> = Default::default();
-        for v in &feature_def.props {
-            let path = format!("features/{}/{}", feature_def.name, v.name);
-            if v.pref_key.is_some() && !v.typ.supports_prefs() {
-                return Err(FMLError::ValidationError(
-                    path,
-                    "Pref keys can only be used with Boolean, String, Int and Text variables"
-                        .to_string(),
-                ));
-            }
-
-            if let Some(sa) = &v.string_alias {
-                let t = &v.typ;
-                let types = t.all_types();
-                if !types.contains(sa) {
-                    return Err(FMLError::ValidationError(
-                        path,
-                        format!(
-                            "The string-alias {sa} must be part of the {} type declaration",
-                            v.name
-                        ),
-                    ));
-                }
-
-                if !string_aliases.insert(sa) {
-                    return Err(FMLError::ValidationError(
-                        path,
-                        format!("The string-alias {sa} should only be declared once per feature"),
-                    ));
-                }
-            }
-        }
-
-        let types = feature_def.all_types();
-        let nm = &feature_def.name;
-        self.validate_string_alias_declarations(
-            &format!("features/{nm}"),
-            nm,
-            &types,
-            &string_aliases,
-        )?;
-
-        Ok(())
-    }
-
-    fn validate_string_alias_declarations(
-        &self,
-        path: &str,
-        feature: &str,
-        types: &HashSet<TypeRef>,
-        string_aliases: &HashSet<&TypeRef>,
-    ) -> Result<()> {
-        let unaccounted: Vec<_> = types
-            .iter()
-            .filter(|t| matches!(t, TypeRef::StringAlias(_)))
-            .filter(|t| !string_aliases.contains(t))
-            .collect();
-
-        if !unaccounted.is_empty() {
-            let t = unaccounted.get(0).unwrap();
-            return Err(FMLError::ValidationError(
-                path.to_string(),
-                format!("A string-alias {t} is used by– but has not been defined in– the {feature} feature"),
-            ));
-        }
-        for t in types {
-            if let TypeRef::Object(nm) = t {
-                if let Some(obj) = self.find_object(nm) {
-                    let types = obj.all_types();
-                    self.validate_string_alias_declarations(
-                        &format!("objects/{nm}"),
-                        feature,
-                        &types,
-                        string_aliases,
-                    )?;
-                }
-            }
         }
         Ok(())
     }
@@ -918,26 +698,6 @@ pub mod unit_tests {
     }
 
     #[test]
-    fn validate_duplicate_enum_defs_fail() -> Result<()> {
-        let mut fm = get_simple_homescreen_feature();
-        fm.enum_defs.insert(
-            "HomeScreenSection".into(),
-            EnumDef {
-                name: "HomeScreenSection".into(),
-                doc: "The sections of the homescreen".into(),
-                variants: vec![
-                    VariantDef::new("top-sites", "The original frecency sorted sites"),
-                    VariantDef::new("jump-back-in", "Jump back in section"),
-                    VariantDef::new("recently-saved", "Tabs that have been bookmarked recently"),
-                ],
-            },
-        );
-        fm.validate_manifest()
-            .expect_err("Should fail on duplicate enum_defs");
-        Ok(())
-    }
-
-    #[test]
     fn validate_allow_coenrollment() -> Result<()> {
         let mut fm = get_simple_homescreen_feature();
         fm.add_feature(FeatureDef::new(
@@ -954,232 +714,6 @@ pub mod unit_tests {
         let coenrolling_ids = fm.get_coenrolling_feature_ids();
         assert_eq!(coenrolling_ids, vec!["some_def".to_string()]);
 
-        Ok(())
-    }
-}
-
-#[cfg(test)]
-mod manifest_structure {
-    use super::*;
-
-    use serde_json::json;
-
-    use crate::fixtures::intermediate_representation::get_simple_homescreen_feature;
-
-    #[test]
-    fn validate_duplicate_feature_defs_fails() -> Result<()> {
-        let mut fm = get_simple_homescreen_feature();
-        fm.add_feature(FeatureDef::new(
-            "homescreen",
-            "Represents the homescreen feature",
-            vec![PropDef::new(
-                "sections-enabled",
-                &TypeRef::EnumMap(
-                    Box::new(TypeRef::Enum("SectionId".into())),
-                    Box::new(TypeRef::String),
-                ),
-                &json!({
-                    "top-sites": true,
-                    "jump-back-in": false,
-                    "recently-saved": false,
-                }),
-            )],
-            false,
-        ));
-        fm.validate_manifest()
-            .expect_err("Should fail on duplicate feature defs");
-        Ok(())
-    }
-
-    #[test]
-    fn validate_duplicate_props_in_same_feature_fails() -> Result<()> {
-        let mut fm = get_simple_homescreen_feature();
-        fm.add_feature(FeatureDef::new(
-            "otherhomescreen",
-            "Represents the homescreen feature",
-            vec![
-                PropDef::new(
-                    "duplicate-prop",
-                    &TypeRef::EnumMap(
-                        Box::new(TypeRef::Enum("SectionId".into())),
-                        Box::new(TypeRef::String),
-                    ),
-                    &json!({
-                        "top-sites": true,
-                        "jump-back-in": false,
-                        "recently-saved": false,
-                    }),
-                ),
-                PropDef::new(
-                    "duplicate-prop",
-                    &TypeRef::EnumMap(
-                        Box::new(TypeRef::Enum("SectionId".into())),
-                        Box::new(TypeRef::String),
-                    ),
-                    &json!({
-                        "top-sites": true,
-                        "jump-back-in": false,
-                        "recently-saved": false,
-                    }),
-                ),
-            ],
-            false,
-        ));
-        fm.validate_manifest()
-            .expect_err("Should fail on duplicate props in the same feature");
-        Ok(())
-    }
-
-    #[test]
-    fn validate_enum_type_ref_doesnt_match_def() -> Result<()> {
-        let mut fm = get_simple_homescreen_feature();
-        fm.add_feature(FeatureDef::new(
-            "some_def",
-            "test doc",
-            vec![PropDef::new(
-                "prop name",
-                &TypeRef::Enum("EnumDoesntExist".into()),
-                &json!(null),
-            )],
-            false,
-        ));
-        fm.validate_manifest().expect_err(
-            "Should fail since EnumDoesntExist isn't a an enum defined in the manifest",
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn validate_obj_type_ref_doesnt_match_def() -> Result<()> {
-        let mut fm = get_simple_homescreen_feature();
-        fm.add_feature(FeatureDef::new(
-            "some_def",
-            "test doc",
-            vec![PropDef::new(
-                "prop name",
-                &TypeRef::Object("ObjDoesntExist".into()),
-                &json!(null),
-            )],
-            false,
-        ));
-        fm.validate_manifest().expect_err(
-            "Should fail since ObjDoesntExist isn't a an Object defined in the manifest",
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn validate_enum_map_with_non_enum_key() -> Result<()> {
-        let mut fm = get_simple_homescreen_feature();
-        fm.add_feature(FeatureDef::new(
-            "some_def",
-            "test doc",
-            vec![PropDef::new(
-                "prop name",
-                &TypeRef::EnumMap(Box::new(TypeRef::String), Box::new(TypeRef::String)),
-                &json!(null),
-            )],
-            false,
-        ));
-        fm.validate_manifest()
-            .expect_err("Should fail since the key on an EnumMap must be an Enum");
-        Ok(())
-    }
-
-    #[test]
-    fn validate_list_with_enum_with_no_def() -> Result<()> {
-        let mut fm = get_simple_homescreen_feature();
-        fm.add_feature(FeatureDef::new(
-            "some_def",
-            "test doc",
-            vec![PropDef::new(
-                "prop name",
-                &TypeRef::List(Box::new(TypeRef::Enum("EnumDoesntExist".into()))),
-                &json!(null),
-            )],
-            false,
-        ));
-        fm.validate_manifest()
-            .expect_err("Should fail EnumDoesntExist isn't a an enum defined in the manifest");
-        Ok(())
-    }
-
-    #[test]
-    fn validate_enum_map_with_enum_with_no_def() -> Result<()> {
-        let mut fm = get_simple_homescreen_feature();
-        fm.add_feature(FeatureDef::new(
-            "some_def",
-            "test doc",
-            vec![PropDef::new(
-                "prop name",
-                &TypeRef::EnumMap(
-                    Box::new(TypeRef::Enum("EnumDoesntExist".into())),
-                    Box::new(TypeRef::String),
-                ),
-                &json!(null),
-            )],
-            false,
-        ));
-        fm.validate_manifest().expect_err(
-            "Should fail since EnumDoesntExist isn't a an enum defined in the manifest",
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn validate_enum_map_with_obj_value_no_def() -> Result<()> {
-        let mut fm = get_simple_homescreen_feature();
-        fm.add_feature(FeatureDef::new(
-            "some_def",
-            "test doc",
-            vec![PropDef::new(
-                "prop name",
-                &TypeRef::EnumMap(
-                    Box::new(TypeRef::Enum("SectionId".into())),
-                    Box::new(TypeRef::Object("ObjDoesntExist".into())),
-                ),
-                &json!(null),
-            )],
-            false,
-        ));
-        fm.validate_manifest()
-            .expect_err("Should fail since ObjDoesntExist isn't an Object defined in the manifest");
-        Ok(())
-    }
-
-    #[test]
-    fn validate_string_map_with_enum_value_no_def() -> Result<()> {
-        let mut fm = get_simple_homescreen_feature();
-        fm.add_feature(FeatureDef::new(
-            "some_def",
-            "test doc",
-            vec![PropDef::new(
-                "prop name",
-                &TypeRef::StringMap(Box::new(TypeRef::Enum("EnumDoesntExist".into()))),
-                &json!(null),
-            )],
-            false,
-        ));
-        fm.validate_manifest()
-            .expect_err("Should fail since ObjDoesntExist isn't an Object defined in the manifest");
-        Ok(())
-    }
-
-    #[test]
-    fn validate_nested_optionals_fail() -> Result<()> {
-        let mut fm = get_simple_homescreen_feature();
-        fm.add_feature(FeatureDef::new(
-            "some_def",
-            "test doc",
-            vec![PropDef::new(
-                "prop name",
-                &TypeRef::Option(Box::new(TypeRef::Option(Box::new(TypeRef::String)))),
-                &json!(null),
-            )],
-            false,
-        ));
-        fm.validate_manifest()
-            .expect_err("Should fail since we can't have nested optionals");
         Ok(())
     }
 }
@@ -1587,110 +1121,6 @@ mod feature_config_tests {
             result.err().unwrap().to_string(),
             "Validation Error at features/feature.prop_1#SampleObj: Invalid key \"invalid-prop\" for object SampleObj; did you mean \"string\"?"
         );
-
-        Ok(())
-    }
-}
-
-#[cfg(test)]
-mod string_aliases {
-    use serde_json::json;
-
-    use super::*;
-
-    fn with_objects(props: &[PropDef], objects: &[ObjectDef]) -> FeatureManifest {
-        let mut fm = with_feature(props);
-        let mut obj_defs: BTreeMap<_, _> = Default::default();
-        for o in objects {
-            obj_defs.insert(o.name(), o.clone());
-        }
-        fm.obj_defs = obj_defs;
-        fm
-    }
-
-    fn with_feature(props: &[PropDef]) -> FeatureManifest {
-        let mut fm: FeatureManifest = Default::default();
-        let f = FeatureDef::new("test-feature", "", props.into(), false);
-        fm.add_feature(f);
-        fm
-    }
-
-    #[test]
-    fn test_validate_feature_structure() -> Result<()> {
-        let name = TypeRef::StringAlias("PersonName".to_string());
-        let all_names = {
-            let t = TypeRef::List(Box::new(name.clone()));
-            let v = json!(["Alice", "Bonnie", "Charlie", "Denise", "Elise", "Frankie"]);
-            PropDef::with_string_alias("all-names", &t, &v, &name)
-        };
-
-        let all_names2 = {
-            let t = TypeRef::List(Box::new(name.clone()));
-            let v = json!(["Alice", "Bonnie"]);
-            PropDef::with_string_alias("all-names-duplicate", &t, &v, &name)
-        };
-
-        // -> Verify that only one property per feature can define the same string-alias.
-        let fm = with_feature(&[all_names.clone(), all_names2.clone()]);
-        assert!(fm.validate_defaults().is_err());
-
-        let newest_member = {
-            let t = &name;
-            let v = json!("Alice"); // it doesn't matter for this test what the value is.
-            PropDef::new("newest-member", t, &v)
-        };
-
-        // -> Verify that a property in a feature can validate against the a string-alias
-        // -> in the same feature.
-        // { all-names: ["Alice"], newest-member: "Alice" }
-        let fm = with_feature(&[all_names.clone(), newest_member.clone()]);
-        fm.validate_defaults()?;
-
-        // { newest-member: "Alice" }
-        // We have a reference to a team mate, but no definitions.
-        // Should error out.
-        let fm = with_feature(&[newest_member.clone()]);
-        assert!(fm.validate_defaults().is_err());
-
-        // -> Validate a property in a nested object can validate against a string-alias
-        // -> in a feature that uses the object.
-        let team_def = ObjectDef::new("Team", &[newest_member.clone()]);
-        let team = {
-            let t = TypeRef::Object("Team".to_string());
-            let v = json!({ "newest-member": "Alice" });
-
-            PropDef::new("team", &t, &v)
-        };
-
-        // { all-names: ["Alice"], team: { newest-member: "Alice" } }
-        let fm = with_objects(&[all_names.clone(), team.clone()], &[team_def.clone()]);
-        fm.validate_defaults()?;
-
-        // { team: { newest-member: "Alice" } }
-        let fm = with_objects(&[team.clone()], &[team_def.clone()]);
-        assert!(fm.validate_defaults().is_err());
-
-        // -> Validate a property in a deeply nested object can validate against a string-alias
-        // -> in a feature that uses the object.
-
-        let match_def = ObjectDef::new("Match", &[team.clone()]);
-        let match_ = {
-            let t = TypeRef::Object("Match".to_string());
-            let v = json!({ "team": { "newest-member": "Alice" }});
-
-            PropDef::new("match", &t, &v)
-        };
-
-        // { all-names: ["Alice"], match: { team: { newest-member: "Alice" }} }
-        let fm = with_objects(
-            &[all_names.clone(), match_.clone()],
-            &[team_def.clone(), match_def.clone()],
-        );
-        fm.validate_defaults()?;
-
-        // { match: {team: { newest-member: "Alice" }} }
-        let fm = with_objects(&[match_.clone()], &[team_def.clone(), match_def.clone()]);
-        assert!(fm.validate_defaults().is_err());
 
         Ok(())
     }
