+++ components/support/nimbus-fml/src/lib.rs
@@ -9,6 +9,7 @@ pub mod error;
 pub(crate) mod frontend;
 pub mod intermediate_representation;
 pub mod parser;
+pub(crate) mod structure;
 pub mod util;
 
 cfg_if::cfg_if! {
